workflow/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
lxml==4.6.3
numpy==1.20.1
opencv-python==4.5.1.48
pathlib==1.0.1
Pillow==8.1.2
reportlab==3.5.66
soupsieve==2.2.1
tesserocr==2.5.1
Click==7.0
